+++ src/main/java/lime1st/netty/service/common/Router.java
@@ -5,8 +5,8 @@ import lime1st.netty.infra.redis.RedisService;
 import lime1st.netty.service.auth.TokenExpire;
 import lime1st.netty.service.auth.TokenIssue;
 import lime1st.netty.service.auth.TokenVerify;
-import lime1st.netty.service.user.UserInfo;
-import lime1st.netty.domain.UserRepository;
+import lime1st.netty.service.user.UserEndpoint;
+import lime1st.netty.user.adapter.out.persistence.UserRepository;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -25,7 +25,7 @@ public class Router {
         routes.put("/tokens_GET", reqData -> new TokenVerify(reqData, redisService));
         routes.put("/tokens_POST", reqData -> new TokenIssue(reqData, userRepository, redisService));
         routes.put("/tokens_DELETE", reqData -> new TokenExpire(reqData, redisService));
-        routes.put("/users_GET", reqData -> new UserInfo(reqData, userRepository));
+        routes.put("/users_GET", reqData -> new UserEndpoint(reqData, userRepository));
     }
 
     public ApiRequest route(String uri, String method, Map<String, String> reqData) {
